# common-platform-rule-engine-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-rule-engine
  description: Rule Engine Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/rules
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/rules/evaluate/plain:
    post:
      tags:
        - Rules Evaluation
      summary: Evaluate plain YAML rules definition
      description: Evaluate a plain (non-base64) YAML DSL rules definition against provided input data. Input variables must use camelCase (creditScore, annualIncome), computed variables will be in snake_case (debt_to_income, is_eligible), and constants use UPPER_CASE (MIN_CREDIT_SCORE). This endpoint accepts YAML content directly without base64 encoding.
      operationId: evaluateRulesPlain
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlainYamlEvaluationRequestDTO'
        required: true
      responses:
        '200':
          description: Rules evaluated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
        '400':
          description: Invalid rules definition, input data, or naming convention violations
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
  /api/v1/rules/evaluate/direct:
    post:
      tags:
        - Rules Evaluation
      summary: Evaluate base64-encoded YAML rules definition
      description: Evaluate a base64-encoded YAML DSL rules definition against provided input data. Input variables must use camelCase (creditScore, annualIncome), computed variables will be in snake_case (debt_to_income, is_eligible), and constants use UPPER_CASE (MIN_CREDIT_SCORE).
      operationId: evaluateRulesDirect
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RulesEvaluationRequestDTO'
        required: true
      responses:
        '200':
          description: Rules evaluated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
        '400':
          description: Invalid rules definition, input data, or naming convention violations
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
  /api/v1/rules/evaluate/by-code:
    post:
      tags:
        - Rules Evaluation
      summary: Evaluate stored rule definition by code
      description: Evaluate a stored YAML DSL rule definition by its code against provided input data. Input variables must use camelCase naming convention. Returns computed variables in snake_case format.
      operationId: evaluateRuleByCode
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleEvaluationByCodeRequestDTO'
        required: true
      responses:
        '200':
          description: Rules evaluated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
        '400':
          description: Invalid input data or naming convention violations
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
        '404':
          description: Rule definition not found or inactive
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RulesEvaluationResponseDTO'
...
components:
  schemas:
...
    PlainYamlEvaluationRequestDTO:
      required:
        - inputData
        - yamlContent
      type: object
      properties:
        yamlContent:
          maxLength: 50000
          minLength: 10
          type: string
          description: 'Plain YAML DSL rules definition to evaluate. Must follow naming conventions: camelCase for inputs (creditScore, annualIncome), snake_case for computed variables (debt_to_income, is_eligible), UPPER_CASE for constants (MIN_CREDIT_SCORE).'
          example: |-
            name: "Credit Scoring Rule"
            description: "Basic credit assessment for loan applications"

            inputs:
              - creditScore        # camelCase input
              - annualIncome       # camelCase input
              - employmentYears    # camelCase input
              - existingDebt       # camelCase input

            when:
              - creditScore at_least MIN_CREDIT_SCORE
              - annualIncome at_least 50000
              - employmentYears at_least 2

            then:
              - calculate debt_to_income as existingDebt / annualIncome    # snake_case computed
              - set is_eligible to true                                   # snake_case computed
              - set approval_tier to "STANDARD"                          # snake_case computed

            else:
              - set is_eligible to false
              - set approval_tier to "DECLINED"

            output:
              is_eligible: is_eligible
              approval_tier: approval_tier
              debt_to_income: debt_to_income
        inputData:
          type: object
          additionalProperties:
            type: object
            description: Input data for rule evaluation. Variable names must follow camelCase convention (e.g., creditScore, annualIncome, employmentYears). These are dynamic values passed from your application.
            example:
              creditScore: 780
              annualIncome: 75000
              employmentYears: 3
              existingDebt: 25000
          description: Input data for rule evaluation. Variable names must follow camelCase convention (e.g., creditScore, annualIncome, employmentYears). These are dynamic values passed from your application.
          example:
            creditScore: 780
            annualIncome: 75000
            employmentYears: 3
            existingDebt: 25000
        metadata:
          type: object
          additionalProperties:
            type: object
            description: Optional metadata for the evaluation request. Use snake_case for metadata keys.
            example:
              request_id: req-123
              source_system: loan_application
              user_id: user-456
              session_id: sess-789
          description: Optional metadata for the evaluation request. Use snake_case for metadata keys.
          example:
            request_id: req-123
            source_system: loan_application
            user_id: user-456
            session_id: sess-789
        includeDetails:
          type: boolean
          description: Whether to include detailed execution information
          example: false
        debugMode:
          type: boolean
          description: Whether to enable debug mode for detailed logging
          example: false
      description: Evaluation request with plain YAML rules definition and camelCase input data
    RulesEvaluationResponseDTO:
      required:
        - success
      type: object
      properties:
        success:
          type: boolean
          description: Whether the rules evaluation was successful
          example: true
        conditionResult:
          type: boolean
          description: The result of the main condition evaluation
          example: true
        outputData:
          type: object
          additionalProperties:
            type: object
            description: Output data generated by the rules. Contains computed variables in snake_case format (e.g., debt_to_income, credit_tier, is_eligible).
            example:
              credit_tier: PRIME
              is_eligible: true
              debt_to_income: 0.33
              risk_score: 85
              approval_status: APPROVED
          description: Output data generated by the rules. Contains computed variables in snake_case format (e.g., debt_to_income, credit_tier, is_eligible).
          example:
            credit_tier: PRIME
            is_eligible: true
            debt_to_income: 0.33
            risk_score: 85
            approval_status: APPROVED
        circuitBreakerTriggered:
          type: boolean
          description: Whether the circuit breaker was triggered
          example: false
        circuitBreakerMessage:
          maxLength: 500
          minLength: 0
          type: string
          description: Circuit breaker message if triggered
          example: High risk transaction detected
        error:
          maxLength: 1000
          minLength: 0
          type: string
          description: Error message if evaluation failed
          example: Variable 'customer_age' not found
        executionTimeMs:
          minimum: 0
          type: integer
          description: Execution time in milliseconds
          format: int64
          example: 25
        metadata:
          type: object
          additionalProperties:
            type: object
            description: Additional metadata about the evaluation
            example:
              rulesId: rules-123
              rulesName: Premium Discount
          description: Additional metadata about the evaluation
          example:
            rulesId: rules-123
            rulesName: Premium Discount
      description: Response from rules evaluation containing results and output data
    RulesEvaluationRequestDTO:
      required:
        - inputData
      type: object
      properties:
        rulesDefinitionBase64:
          maxLength: 100000
          minLength: 0
          type: string
          description: 'Base64 encoded YAML rules definition to evaluate. Must follow naming conventions: camelCase for inputs, snake_case for computed variables, UPPER_CASE for constants.'
          example: bmFtZTogIkNyZWRpdCBBc3Nlc3NtZW50IgpkZXNjcmlwdGlvbjogIkJhc2ljIGNyZWRpdCBlbGlnaWJpbGl0eSBjaGVjayIKCmlucHV0czoKICAtIGNyZWRpdFNjb3JlICAgICAgIyBjYW1lbENhc2UgaW5wdXQKICAtIGFubual SW5jb21lICAgICAgIyBjYW1lbENhc2UgaW5wdXQKCndoZW46CiAgLSBjcmVkaXRTY29yZSBhdF9sZWFzdCBNSU5fQ1JFRElUX1NDT1JFICMgVVBQRVJfQ0FTRSBjb25zdGFudAogIC0gYW5udWFsSW5jb21lIGF0X2xlYXN0IDUwMDAwCgp0aGVuOgogIC0gY2FsY3VsYXRlIGRlYnRfdG9faW5jb21lIGFzIGV4aXN0aW5nRGVidCAvIGFubual SW5jb21lICMgc25ha2VfY2FzZSBjb21wdXRlZAogIC0gc2V0IGNyZWRpdF90aWVyIHRvICJQUklNRSIgICMgc25ha2VfY2FzZSBjb21wdXRlZAogIC0gc2V0IGlzX2VsaWdpYmxlIHRvIHRydWUgICMgc25ha2VfY2FzZSBjb21wdXRlZAoKZWxzZToKICAtIHNldCBjcmVkaXRfdGllciB0byAiU1RBTkRBUkQiCiAgLSBzZXQgaXNfZWxpZ2libGUgdG8gZmFsc2UKCm91dHB1dDoKICBjcmVkaXRfdGllcjogY3JlZGl0X3RpZXIKICBpc19lbGlnaWJsZTogaXNfZWxpZ2libGUKICBkZWJ0X3RvX2luY29tZTogZGVidF90b19pbmNvbWU=
        inputData:
          type: object
          additionalProperties:
            type: object
            description: Input data for rule evaluation. Variable names must follow camelCase convention (e.g., creditScore, annualIncome, employmentYears). These are dynamic values passed from your application.
            example:
              creditScore: 780
              annualIncome: 75000
              employmentYears: 3
              existingDebt: 25000
              requestedAmount: 200000
          description: Input data for rule evaluation. Variable names must follow camelCase convention (e.g., creditScore, annualIncome, employmentYears). These are dynamic values passed from your application.
          example:
            creditScore: 780
            annualIncome: 75000
            employmentYears: 3
            existingDebt: 25000
            requestedAmount: 200000
        metadata:
          type: object
          additionalProperties:
            type: object
            description: Optional metadata for the evaluation request. Use snake_case for metadata keys.
            example:
              request_id: req-123
              source_system: loan_application
              user_id: user-456
              session_id: sess-789
          description: Optional metadata for the evaluation request. Use snake_case for metadata keys.
          example:
            request_id: req-123
            source_system: loan_application
            user_id: user-456
            session_id: sess-789
        includeDetails:
          type: boolean
          description: Whether to include detailed execution information
          example: false
        debugMode:
          type: boolean
          description: Whether to enable debug mode for detailed logging
          example: false
      description: Evaluation request with base64-encoded YAML rules definition and camelCase input data
    RuleEvaluationByCodeRequestDTO:
      required:
        - inputData
        - ruleDefinitionCode
      type: object
      properties:
        ruleDefinitionCode:
          pattern: ^[a-zA-Z][a-zA-Z0-9_]*$
          type: string
          description: Code of the stored rule definition to evaluate
          example: credit_scoring_v1
        inputData:
          type: object
          additionalProperties:
            type: object
            description: Input data for rule evaluation. Variable names must follow camelCase convention (e.g., creditScore, annualIncome, employmentYears). These are dynamic values passed from your application.
            example:
              creditScore: 780
              annualIncome: 75000
              employmentYears: 3
              existingDebt: 25000
              requestedAmount: 200000
          description: Input data for rule evaluation. Variable names must follow camelCase convention (e.g., creditScore, annualIncome, employmentYears). These are dynamic values passed from your application.
          example:
            creditScore: 780
            annualIncome: 75000
            employmentYears: 3
            existingDebt: 25000
            requestedAmount: 200000
        metadata:
          type: object
          additionalProperties:
            type: object
            description: Optional metadata for the evaluation request. Use snake_case for metadata keys.
            example:
              request_id: req-123
              source_system: loan_application
              user_id: user-456
              session_id: sess-789
          description: Optional metadata for the evaluation request. Use snake_case for metadata keys.
          example:
            request_id: req-123
            source_system: loan_application
            user_id: user-456
            session_id: sess-789
        includeDetails:
          type: boolean
          description: Whether to include detailed execution information
          example: false
        debugMode:
          type: boolean
          description: Whether to enable debug mode for detailed logging
          example: false
      description: Evaluation request with rule code and camelCase input data
...